Moves LastUpdatedOrderedDict to common.py
